Multiple Updates - TLSA Record Changes - Code Removal
This commit removed unnecessary code at the bottom of this script that wasn't doing anything.

This also changes the way TLSA records are generated. Instead of getting a hash of the whole cert ( 3 0 1 prefix) we are now getting the hash for subject public key ( 3 1 1 Prefix)

This also now generates a file and saves the TLSA record in it so that the logs isn't the only place it shows up.